# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "🌸 Anime Garden API",
    "description": "🌸 Anime Garden 是 [动漫花园](https://share.dmhy.org/) 第三方镜像站以及动画 BT 资源聚合站。\n\n+ ☁️ 为开发者准备的开放 API 接口\n+ 📺 查看动画放送时间表来找到你喜欢的动画\n+ 🔖 支持丰富的高级搜索，例如：`葬送的芙莉莲 +简体内嵌 字幕组:桜都字幕组 类型:动画`\n+ 📙 自定义 RSS 订阅链接，例如：[葬送的芙莉莲](https://api.animes.garden/feed.xml?filter=%5B%7B%22fansubId%22:%5B%22619%22%5D,%22type%22:%22%E5%8B%95%E7%95%AB%22,%22include%22:%5B%22%E8%91%AC%E9%80%81%E7%9A%84%E8%8A%99%E8%8E%89%E8%8E%B2%22%5D,%22keywords%22:%5B%22%E7%AE%80%E4%BD%93%E5%86%85%E5%B5%8C%22%5D%7D%5D)\n+ ⭐ 搜索条件收藏夹和生成聚合的 RSS 订阅链接\n+ 👷‍♂️ 支持与 AutoBangumi 和 AnimeSpace 集成\n",
    "version": "0.5.4",
    "license": {
      "name": "",
      "url": "https://github.com/yjl9903/AnimeGarden/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:9701",
      "description": "本地 API 服务器 (dmbt-web)"
    }
  ],
  "paths": {
...
    "/resources": {
      "get": {
        "summary": "获取资源列表",
        "description": "获取所有提供商的资源列表，支持分页和过滤",
        "operationId": "getResources",
        "tags": [
          "Resources"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/PageSizeParam"
          },
          {
            "$ref": "#/components/parameters/ProviderParam"
          },
          {
            "$ref": "#/components/parameters/SearchParam"
          },
          {
            "$ref": "#/components/parameters/IncludeParam"
          },
          {
            "$ref": "#/components/parameters/KeywordsParam"
          },
          {
            "$ref": "#/components/parameters/ExcludeParam"
          },
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/SubjectParam"
          },
          {
            "$ref": "#/components/parameters/FansubParam"
          },
          {
            "$ref": "#/components/parameters/PublisherParam"
          },
          {
            "$ref": "#/components/parameters/AfterParam"
          },
          {
            "$ref": "#/components/parameters/BeforeParam"
          },
          {
            "$ref": "#/components/parameters/DuplicateParam"
          },
          {
            "$ref": "#/components/parameters/TrackerParam"
          },
          {
            "$ref": "#/components/parameters/MetadataParam"
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取资源列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "获取资源列表 (POST)",
        "description": "使用POST方法获取资源列表，支持复杂的过滤条件",
        "operationId": "postResources",
        "tags": [
          "Resources"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TrackerParam"
          },
          {
            "$ref": "#/components/parameters/MetadataParam"
          }
        ],
        "requestBody": {
          "description": "过滤和分页参数",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourcesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功获取资源列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcesResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/collection": {
      "post": {
        "summary": "创建收藏夹",
        "description": "创建新的收藏夹",
        "operationId": "createCollection",
        "tags": [
          "Collections"
        ],
        "requestBody": {
          "required": true,
          "description": "收藏夹信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建收藏夹",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "id": {
                      "type": "integer",
                      "description": "收藏夹ID"
                    },
                    "hash": {
                      "type": "string",
                      "description": "收藏夹哈希值"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "status",
                    "id",
                    "hash",
                    "createdAt",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "收藏夹格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "创建或更新收藏夹",
        "description": "创建或更新收藏夹（与POST相同）",
        "operationId": "upsertCollection",
        "tags": [
          "Collections"
        ],
        "requestBody": {
          "required": true,
          "description": "收藏夹信息",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建收藏夹",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "OK"
                      ]
                    },
                    "id": {
                      "type": "integer",
                      "description": "收藏夹ID"
                    },
                    "hash": {
                      "type": "string",
                      "description": "收藏夹哈希值"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "status",
                    "id",
                    "hash",
                    "createdAt",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "收藏夹格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/feed.xml": {
      "get": {
        "summary": "获取RSS订阅源",
        "description": "获取资源的RSS订阅源",
        "operationId": "getRSSFeed",
        "tags": [
          "Feed"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchParam"
          },
          {
            "$ref": "#/components/parameters/IncludeParam"
          },
          {
            "$ref": "#/components/parameters/KeywordsParam"
          },
          {
            "$ref": "#/components/parameters/ExcludeParam"
          },
          {
            "$ref": "#/components/parameters/TypeParam"
          },
          {
            "$ref": "#/components/parameters/SubjectParam"
          },
          {
            "$ref": "#/components/parameters/FansubParam"
          },
          {
            "$ref": "#/components/parameters/PublisherParam"
          },
          {
            "$ref": "#/components/parameters/AfterParam"
          },
          {
            "$ref": "#/components/parameters/BeforeParam"
          },
          {
            "name": "tracker",
            "in": "query",
            "description": "是否包含Tracker信息",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "on",
                "off",
                "yes",
                "no"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "RSS XML内容",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string",
                  "description": "RSS XML格式的订阅源"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "页码",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSizeParam": {
        "name": "pageSize",
        "in": "query",
        "description": "每页数量",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 100
        }
      },
      "ProviderParam": {
        "name": "provider",
        "in": "query",
        "description": "提供商过滤",
        "schema": {
          "type": "string",
          "enum": [
            "dmhy",
            "mikan",
            "moe",
            "ani"
          ]
        }
      },
      "SearchParam": {
        "name": "search",
        "in": "query",
        "description": "搜索关键词（可重复传递多个参数，例如 ?search=foo&search=bar，或在Swagger页面中以逗号分隔输入：foo,bar）",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "IncludeParam": {
        "name": "include",
        "in": "query",
        "description": "包含关键词（可重复传递多个参数，例如 ?include=foo&include=bar，或在Swagger页面中以逗号分隔输入：foo,bar）",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "KeywordsParam": {
        "name": "keywords",
        "in": "query",
        "description": "必须包含的关键词",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "ExcludeParam": {
        "name": "exclude",
        "in": "query",
        "description": "排除关键词",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "TypeParam": {
        "name": "type",
        "in": "query",
        "description": "资源类型过滤",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "SubjectParam": {
        "name": "subject",
        "in": "query",
        "description": "Bangumi ID 过滤",
        "schema": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "style": "form",
        "explode": true
      },
      "FansubParam": {
        "name": "fansub",
        "in": "query",
        "description": "字幕组过滤",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "PublisherParam": {
        "name": "publisher",
        "in": "query",
        "description": "发布者过滤",
        "schema": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "style": "form",
        "explode": true
      },
      "AfterParam": {
        "name": "after",
        "in": "query",
        "description": "起始时间过滤（ISO 8601格式）",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "BeforeParam": {
        "name": "before",
        "in": "query",
        "description": "结束时间过滤（ISO 8601格式）",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "DuplicateParam": {
        "name": "duplicate",
        "in": "query",
        "description": "是否包含重复资源",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "TrackerParam": {
        "name": "tracker",
        "in": "query",
        "description": "是否返回tracker信息",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "MetadataParam": {
        "name": "metadata",
        "in": "query",
        "description": "是否返回metadata信息",
        "schema": {
          "type": "boolean",
          "default": false
        }
      }
    },
    "schemas": {
...
      "Resource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "provider": {
            "type": "string"
          },
          "providerId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "href": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "magnet": {
            "type": "string"
          },
          "tracker": {
            "type": "string",
            "nullable": true,
            "description": "当tracker参数为true时返回"
          },
          "size": {
            "type": "integer"
          },
          "fansub": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name"
            ]
          },
          "publisher": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "name": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name"
            ]
          },
          "subjectId": {
            "type": "integer",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "fetchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "description": "当metadata参数为true时返回",
            "properties": {
              "anipar": {
                "type": "object",
                "description": "Anipar解析结果"
              }
            }
          }
        },
        "required": [
          "id",
          "provider",
          "providerId",
          "title",
          "href",
          "type",
          "magnet",
          "size",
          "publisher",
          "createdAt",
          "fetchedAt"
        ]
      },
...
      "FilterOptions": {
        "type": "object",
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "bangumi"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "dmhy",
              "mikan",
              "moe",
              "ani"
            ]
          },
          "duplicate": {
            "type": "boolean"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "after": {
            "type": "string",
            "format": "date-time"
          },
          "before": {
            "type": "string",
            "format": "date-time"
          },
          "fansubs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publishers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "search": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "include": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "keywords": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exclude": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "complete": {
            "type": "boolean"
          }
        },
        "required": [
          "page",
          "pageSize",
          "complete"
        ]
      },
      "ResourcesRequest": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 100
          },
          "provider": {
            "type": "string",
            "enum": [
              "dmhy",
              "mikan",
              "moe",
              "ani"
            ]
          },
          "duplicate": {
            "type": "boolean",
            "default": false
          },
          "after": {
            "type": "string",
            "format": "date-time"
          },
          "before": {
            "type": "string",
            "format": "date-time"
          },
          "search": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "include": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "keywords": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "exclude": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "type": {
            "type": "string"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "subject": {
            "type": "integer"
          },
          "subjects": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "fansub": {
            "type": "string"
          },
          "fansubs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publisher": {
            "type": "string"
          },
          "publishers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "preset": {
            "type": "string",
            "enum": [
              "bangumi"
            ]
          }
        }
      },
      "ResourcesResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OK"
            ]
          },
          "complete": {
            "type": "boolean",
            "description": "为了兼容性保留，与pagination.complete相同"
          },
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          },
          "filter": {
            "$ref": "#/components/schemas/FilterOptions"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "status",
          "complete",
          "resources",
          "pagination",
          "filter",
          "timestamp"
        ]
      },
      "CollectionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "authorization": {
            "type": "string"
          },
          "filters": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "searchParams": {
                  "type": "string"
                }
              },
              "required": [
                "name"
              ],
              "additionalProperties": true
            }
          }
        },
        "required": [
          "name",
          "authorization",
          "filters"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ERROR"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ]
      }
    }
  },
...
}